--- src/components/preview-ghost/PreviewGhost.js
@@ -1,13 +1,0 @@
-// @flow
-import * as React from 'react';
-import Ghost from '../ghost/Ghost';
-
-function PreviewGhost() {
-    return (
-        <div className="PreviewGhost">
-            <Ghost className="PreviewGhost-ghost" height={288} />
-        </div>
-    );
-}
-
-export default PreviewGhost;
