+++ src/features/message-center/components/message/Message.js
@@ -2,7 +2,7 @@
 import * as React from 'react';
 import type { Token } from '../../../../common/types/core';
 import type { EligibleMessageCenterMessage } from '../../types';
-import type { ContentPreviewProps } from '../../../../common/types/contentPreview';
+import type { ContentPreviewProps } from '../../../message-preview-content/MessagePreviewContent';
 import { PreviewTitleBodyTags, PreviewTitleBodyTagsButton, TitleBodyTags, TitleBodyTagsButton } from '../templates';
 import {
     PREVIEW_TITLE_BODY_TAGS,
