+++ src/features/message-center/components/templates/preview-title-body-tags-button/PreviewTitleBodyTagsButton.js
@@ -1,13 +1,13 @@
 // @flow
 import * as React from 'react';
 import type { Token } from '../../../../../common/types/core';
-import MessagePreviewContent from '../../../../../components/message-preview-content/MessagePreviewContent';
+import MessagePreviewContent from '../../../../message-preview-content/MessagePreviewContent';
+import type { ContentPreviewProps } from '../../../../message-preview-content/MessagePreviewContent';
 import MessageTextContent from '../common/MessageTextContent';
 import MessageTags from '../common/MessageTags';
 import MessageFooter from '../common/MessageFooter';
 import BottomContentWrapper from '../common/BottomContentWrapper';
 import type { PreviewTitleBodyTagsButtonTemplateParams } from '../../../types';
-import type { ContentPreviewProps } from '../../../../../common/types/contentPreview';
 
 type Props = $Diff<
     {|
