+++ src/features/message-preview-content/messages.js
@@ -2,7 +2,7 @@
 import { defineMessages } from 'react-intl';
 
 const messages = defineMessages({
-    previewError: {
+    messagePreviewError: {
         defaultMessage: `Sorry, we're having trouble showing this image. `,
         description: 'Error message for preview not loading an image',
         id: 'be.messageCenter.previewError',
