+++ src/features/message-preview-content/styles/PreviewErrorNotification.scss
